replacing all the short ints with regular ints in order to bring the API in line with the native engine's and to make it easier to use

+++ ProcyonSharp.Bindings/Drawing/Sprite.cs
@@ -6,15 +6,15 @@ namespace ProcyonSharp.Bindings.Drawing;
 public class Sprite : NativeObject
 {
     // ReSharper disable once SuggestBaseTypeForParameter
-    internal Sprite(SpriteSheet spriteSheet, short x, short y, short width, short height)
+    internal Sprite(SpriteSheet spriteSheet, int x, int y, int width, int height)
     {
         Size = (width, height);
         Offset = (x, y);
         Pointer = CreateSprite(spriteSheet.Pointer, x, y, width, height);
     }
 
-    public (short, short) Size { get; }
-    public (short, short) Offset { get; }
+    public (int, int) Size { get; }
+    public (int, int) Offset { get; }
 
     protected override void Cleanup()
     {
@@ -22,7 +22,7 @@ public class Sprite : NativeObject
     }
 
     [DllImport("procyon", EntryPoint = "procy_create_sprite")]
-    private static extern IntPtr CreateSprite(IntPtr spriteSheet, short x, short y, short width, short height);
+    private static extern IntPtr CreateSprite(IntPtr spriteSheet, int x, int y, int width, int height);
 
     [DllImport("procyon", EntryPoint = "procy_destroy_sprite")]
     private static extern IntPtr DestroySprite(IntPtr ptr);
